ai-chat-backend/ai-chat-backend/requirements.txt
#FastAPI
fastapi==0.97.0
uvicorn==0.22.0
websockets==11.0.3

#OpenAI
openai==0.27.8
pandas==2.0.2
numpy==1.24.3
matplotlib==3.7.1
plotly==5.15.0
scipy==1.10.1
langdetect==1.0.9
tiktoken==0.4.0

#EnvironmentVariables
python-dotenv==1.0.0

#MongoDB
pymongo==4.3.3

#Keyclok
python-keycloak==3.0.0

#API Rate limiter
slowapi==0.1.8

#Cron Job
apscheduler==3.10.1

#Keycloak
python-keycloak==3.0.0

#HTTP
requests==2.31.0
